autogen_implementations/end_to_end_automotive_damage_analysis_agents/requirements.txt
pyautogen
pillow
python-dotenv
flask
